Refactor(gahee/club-list) : 변수명 변경 및 오류 메세지 변경 #68
